+++ app/forms/consignors/ConsignorEORIFormProvider.scala
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package forms
+package forms.consignors
+
+import forms.mappings.Mappings
+import models.GbEori
+import play.api.data.Form
 
 import javax.inject.Inject
 
-import forms.mappings.Mappings
-import play.api.data.Form
+class ConsignorEORIFormProvider @Inject() extends Mappings {
 
-class ConsignorNameFormProvider @Inject() extends Mappings {
-
-  def apply(): Form[String] =
+  def apply(): Form[GbEori] =
     Form(
-      "value" -> text("consignorName.error.required")
-        .verifying(maxLength(35, "consignorName.error.length"))
+      "value" -> gbEori("consignorEORI.error.required")
     )
 }
